Added leveling and leveling cost displays.  Still need to implement what is GAINED from leveling, but it's there!  (Default SP to 100 for now)

+++ Assets/Game/Scripts/Player Skills/Interfaces/ILevelable.cs
@@ -17,7 +17,10 @@ public interface ILevelable
     /// </summary>
     public byte Rank { get; }
 
-    public ulong LevelCost { get; }
+    /// <summary>
+    /// The cost of going up one level, whatever that entails.
+    /// </summary>
+    public double LevelCost { get; }
 
     public static GrowthType Growth { get; }
